fix changes

+++ server/routes/users.js
@@ -14,7 +14,7 @@ export default (app) => {
       reply.render('users/new', { user });
     })
     .get('/users/:id/edit', { name: 'editUser' }, async (req, reply) => {
-      if (req.user.id === +req.params.id) {
+      if (req.user && req.user.id === +req.params.id) {
         const user = await app.objection.models.user.query().findById(req.params.id);
         reply.render('users/edit', { user });
       } else {
@@ -38,7 +38,7 @@ export default (app) => {
     })
     .patch('/users/:id', { name: 'updateUser' }, async (req, reply) => {
       try {
-        if (req.user.id === +req.params.id) {
+        if (req.user && req.user.id === +req.params.id) {
           const user = await app.objection.models.user.query().findById(+req.params.id);
           await user.$query().patch(req.body.data);
           req.flash('info', i18next.t('flash.users.edit.success'));
@@ -54,14 +54,19 @@ export default (app) => {
         return reply;
       }
     })
-    .delete('/users/:id', async (req, reply) => {
-      if (req.user.id === +req.params.id) {
-        await app.objection.models.user.query().deleteById(+req.params.id);
-        req.flash('info', i18next.t('flash.users.delete.success'));
-      } else {
+    .delete('/users/:id', { name: 'deleteUser' }, async (req, reply) => {
+      try {
+        if (req.user && req.user.id === +req.params.id) {
+          await app.objection.models.user.query().deleteById(+req.params.id);
+          req.flash('info', i18next.t('flash.users.delete.success'));
+        } else {
+          req.flash('info', i18next.t('flash.users.delete.error'));
+        }
+      } catch (er) {
+        console.log(er);
         req.flash('info', i18next.t('flash.users.delete.error'));
       }
-      reply.redirect(app.reverse('root'));
+      reply.redirect(app.reverse('users'));
       return reply;
     });
 };
